Move Pathfinding to separate class
